Setup Algoliasearch PHP client with Composer
Installed Algoliasearch PHP client with Composer and added dynamic import of the library only if we have clicked on the button to send the products to Algolia.

This way we won't include the library on every single Wordpress page.

Remember that we need to do a "composer install" before the plugin will work properly as the "vendor" directory has been added to .gitignore